requirements.txt
anyio==4.12.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@643beff3883b5720d94b6b9c9eca12fa9fb72fb1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.6
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.3
mkl==2021.4.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.9.79
onnx==1.17.0
onnxruntime-gpu==1.22.0
onnxslim==0.1.57
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.3
safetensors==0.7.0
scipy==1.15.3
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
tbb==2021.11.0
tensorrt==10.12.0.36
tensorrt_cu12==10.12.0.36
tensorrt_cu12_bindings==10.12.0.36
tensorrt_cu12_libs==10.12.0.36
timm==1.0.24
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.4.6
ultralytics-thop==2.0.18
urllib3==2.4.0
wcwidth==0.2.14
wheel==0.45.1
